# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Social App API
  description: API documentation for the Social App backend
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/settings/users/mode:
    get:
      tags:
      - Settings
      summary: Get user application mode
      description: Retrieves the authenticated user's current application mode (LIGHT
        or DARK)
      operationId: getMode
      parameters:
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      responses:
        "200":
          description: Application mode retrieved successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid or expired Firebase token
      security:
      - BearerAuth: []
    post:
      tags:
      - Settings
      summary: Update user application mode
      description: Updates the authenticated user's application mode (LIGHT or DARK)
      operationId: setMode
      parameters:
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModeResponse"
        required: true
      responses:
        "200":
          description: Application mode updated successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid or expired Firebase token
        "400":
          description: Invalid mode value
      security:
      - BearerAuth: []
...
  /api/auth/register:
    post:
      tags:
      - Authentication
      summary: Register a new user
      description: Registers a new user using a Firebase ID token and registration
        details
      operationId: registerUser
      parameters:
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: User registered successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid Firebase token
        "409":
          description: User already exists
  /api/auth/logout:
    post:
      tags:
      - Authentication
      summary: Logout current user
      description: Logs the authenticated user out by revoking their Firebase refresh
        tokens
      operationId: logout
      parameters:
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      responses:
        "200":
          description: Logged out successfully
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Invalid Firebase token
  /api/auth/login:
    post:
      tags:
      - Authentication
      summary: Login an existing user
      description: Authenticates an existing user using a Firebase ID token
      operationId: login
      parameters:
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      responses:
        "200":
          description: Login successful
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid Firebase token
        "403":
          description: Account is banned or suspended
        "404":
          description: User not found
  /api/auth/admin/login:
    post:
      tags:
      - Authentication
      summary: Admin login
      description: Authenticates an admin user using a Firebase ID token
      operationId: adminLogin
      parameters:
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      responses:
        "200":
          description: Admin login successful
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid or expired token
        "403":
          description: Not an admin
        "404":
          description: User not found
...
  /api/auth/profile:
    get:
      tags:
      - Authentication
      summary: Get current user profile
      description: Returns the profile of the currently authenticated user
      operationId: getProfile
      responses:
        "200":
          description: Profile retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
components:
  schemas:
    Address:
      type: object
      properties:
        addressid:
          type: integer
          format: int32
        streetNumber:
          type: integer
          format: int32
        street:
          type: string
        zipcode:
          type: integer
          format: int32
        neighbourhoodid:
          $ref: "#/components/schemas/Location"
...
    Badges:
      type: object
      properties:
        badgeid:
          type: integer
          format: int32
        badgeName:
          type: string
        badge_description:
          type: string
        isSpecialist:
          type: boolean
        xpReward:
          type: integer
          format: int32
        ratingid:
          $ref: "#/components/schemas/Ratings"
        badgeDescription:
          type: string
...
    Location:
      type: object
      properties:
        locationid:
          type: integer
          format: int32
        locationCenterPoint:
          type: integer
          format: int32
        locationRadius:
          type: integer
          format: int32
        neighbourhoodid:
          type: integer
          format: int32
        neighbourhoodName:
          type: string
    Ratings:
      type: object
      properties:
        ratingid:
          type: integer
          format: int32
        ratingReview:
          type: string
        totalXpLevel:
          type: integer
          format: int32
        currentGroup:
          type: string
...
    User:
      type: object
      properties:
        userid:
          type: integer
          format: int32
        firebaseUid:
          type: string
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        gender:
          type: string
        addressid:
          $ref: "#/components/schemas/Address"
        badgeid:
          $ref: "#/components/schemas/Badges"
        ratingid:
          $ref: "#/components/schemas/Ratings"
        userType:
          type: string
        isAdmin:
          type: boolean
...
    ModeResponse:
      type: object
      properties:
        mode:
          type: string
...
    RegisterRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        gender:
          type: string
        userType:
          type: string
        username:
          type: string
        addressId:
          type: integer
          format: int32
        badgeId:
          type: integer
          format: int32
        ratingId:
          type: integer
          format: int32
...